Added a blank home page
Added a blank home page with a coming soon message, a non-functional search bar, and an avatar for the test user. Also added and routed temporary buttons to go back and forth between the landing page and home page. Struggling to center the search bar. LOL. Login to the page DOES NOT take you to the home page. For some reason the backend server was giving errors about MongoDB.