Added the Remember Me checkbox functionality to the login page.
If the remember-me checkbox is checked when the submit (sign in) button is pressed, and the credentials pass front-end validation, the email and the checkbox state (true) is saved in the browser's local storage. If the checkbox is later unchecked when login occurs, the variables remain in the local storage, but their values change to: email='' and Remember Me='false'. The password is never stored in the local storage.